pyproject.toml
[project]
name = "agent-litreview"
version = "0.1.0"
description = "Agent-native literature review system. BM25 search, citation PageRank, funnel retrieval, and progressive summarization over 10K+ papers."
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
keywords = ["literature-review", "research", "papers", "agents", "BM25", "citation-graph"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "ruamel.yaml>=0.18.0",
    "requests>=2.31.0",
    "jinja2>=3.1.0",
    "pymupdf>=1.24.0",
]

[project.optional-dependencies]
pagerank = ["scipy>=1.11.0"]
dev = [
    "pytest>=8.0.0",
    "responses>=0.25.0",
]

[project.scripts]
lit = "literature.scripts.lit:main"

[project.urls]
Homepage = "https://github.com/hangyuzhou/agent-litreview"
Repository = "https://github.com/hangyuzhou/agent-litreview"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["literature"]

[tool.uv]
dev-dependencies = ["pytest>=8.0.0", "responses>=0.25.0"]

[tool.pytest.ini_options]
testpaths = ["literature/tests"]
